app/requirements.txt
Flask==3.0.3
mysql-connector-python==9.0.0
requests==2.32.5
pandas==2.2.2